RavenDB-18128 cleanup
