+++ test/Tests.Infrastructure/MultiplatformFactAttribute.cs
@@ -94,27 +94,3 @@ public class MultiplatformFactAttribute : FactAttribute
         }
     }
 }
-
-public class MultiplatformTheoryAttribute : TheoryAttribute
-{
-    private readonly RavenPlatform _platform;
-    private readonly RavenArchitecture _architecture;
-
-    public MultiplatformTheoryAttribute(RavenPlatform platform = RavenPlatform.All)
-        : this(platform, RavenArchitecture.All)
-    {
-    }
-
-    public MultiplatformTheoryAttribute(RavenArchitecture architecture = RavenArchitecture.All)
-        : this(RavenPlatform.All, architecture)
-    {
-    }
-
-    public MultiplatformTheoryAttribute(RavenPlatform platform = RavenPlatform.All, RavenArchitecture architecture = RavenArchitecture.All)
-    {
-        _platform = platform;
-        _architecture = architecture;
-    }
-
-    public override string Skip => MultiplatformFactAttribute.ShouldSkip(_platform, _architecture);
-}
